backend/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
qdrant-client
google-generativeai
requests
